refactor: transaction dialogs
