+++ apps/web/src/components/transactions/TransactionFormFields.tsx
@@ -1,33 +1,33 @@
 "use client";
 
-import { Button } from "@coin-guard/ui";
-import { Calendar } from "@coin-guard/ui";
-import { FieldGroup } from "@coin-guard/ui";
+import { FormType } from "@/constants/forms";
+import { useGetCategories } from "@/hooks/categories/useGetCategories";
+import type {
+  addTransactionSchemaType,
+  editTransactionSchemaType,
+} from "@/schemas/transactions";
 import {
+  Button,
+  Calendar,
+  cn,
+  FieldGroup,
   FormControl,
   FormField,
   FormItem,
   FormLabel,
   FormMessage,
-} from "@coin-guard/ui";
-import { Input } from "@coin-guard/ui";
-import { Popover, PopoverContent, PopoverTrigger } from "@coin-guard/ui";
-import { SearchableSelect } from "@coin-guard/ui";
-import {
+  Input,
+  Popover,
+  PopoverContent,
+  PopoverTrigger,
+  SearchableSelect,
   Select,
   SelectContent,
   SelectItem,
   SelectTrigger,
   SelectValue,
+  Textarea,
 } from "@coin-guard/ui";
-import { Textarea } from "@coin-guard/ui";
-import { FormType } from "@/constants/forms";
-import { useGetCategories } from "@/hooks/categories/useGetCategories";
-import { cn } from "@coin-guard/ui";
-import type {
-  addTransactionSchemaType,
-  editTransactionSchemaType,
-} from "@/schemas/transactions";
 import { format } from "date-fns";
 import { CalendarIcon } from "lucide-react";
 import { useMemo } from "react";
